refactor: Update Schema to BIGINT & Standardize UTC Handling

+++ src/main/java/com/iviet/ivshs/entities/BaseMetricData.java
@@ -26,12 +26,12 @@ public abstract class BaseMetricData extends BaseEntity {
   protected Instant timestamp;
 
   @Column(name = "unix_minute")  
-  protected Integer unixMinute;
+  protected Long unixMinute;
 
   public void setTimestamp(Instant timestamp) {
     this.timestamp = timestamp;
     if (timestamp != null) {
-      this.unixMinute = (int) (timestamp.getEpochSecond() / 60);
+      this.unixMinute = timestamp.getEpochSecond() / 60;
     } else {
       this.unixMinute = null;
     }
